Fix error handling bug, and be more careful with image dimensions.
Improve the image dimension checks, and fix the error message
to not dereference an int as if it was a string.

This solves a crash with MuPDF and samples_mupdf_001/1006.pdf.SIGSEGV.dab.10,
a file supplied by Mateusz "j00ru" Jurczyk and Gynvael Coldwind of the Google
Security Team using Address Sanitizer. Many thanks!