Bug 694880: fix write access violation in opj_j2k_add_tlmarker
A tile_index' current_tpsno may be far larger than the number of tile
parts because opj_j2k_read_sot increases that number just by 10 instead
of growing it to the actually required size.

This can be seen e.g. in:

147af3f1083de4393666b7d99b01b58b_signal_sigsegv_130c531_6155_5136.pdf

Thanks to Mateusz Jurczyk and Gynvael Coldwind of the Google Security
Team for providing the example files.